% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{A numeric value representing the number of people in the group}
}
\value{
The probability (a numeric value between 0 and 1) that at least two people in the group of size x share a birthday
}
\description{
Birthday Function
}
\examples{
birthday(10)
}
