% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Barplot Function}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of trials to run the simulation}

\item{n}{Sample size}

\item{p}{Probability of success in a single Bernoulli trial}
}
\value{
A vector of relative frequencies (empirical probabilities) for each possible number of successes (0 to n), plus a barplot visualization
}
\description{
Binomial Barplot Function
}
\examples{
mybin(100,10,0.5)
}
