% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of bootstrap iterations to perform}

\item{x}{numeric vector containing the sample data}

\item{fun}{the statistic to compute (default is "mean")}

\item{alpha}{significance level for the confidence level}

\item{cx}{size of the text in the plot annotations}

\item{...}{additional graphical parameters passed to 'hist()'}
}
\value{
a list containing the bootstrap confidence interval, the statistic used, the original data vector, and a histogram of the desired statistic
}
\description{
Bootstrap Interval Function
}
\examples{
#set.seed(10); sam(rnorm(20,mean=10,sd=4))
#myboot2(iter=10000,x=sam,fun="mean",alpha-0.05)
}
