% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Function with Starting Uniform Distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size for each iteration}

\item{iter}{number of iterations (sample means)}

\item{a}{lower bound of Uniform distribution}

\item{b}{upper bound of Uniform distribution}
}
\value{
numeric vector of length iter containing the sample means as well as a histogram of the means
}
\description{
Central Limit Theorem Function with Starting Uniform Distribution
}
\examples{
mycltu(n=20, iter=10000, a=0, b=5)
}
