% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper bound for shaded area}
}
\value{
a plot with normal curve and shaded area that goes from negative infinity to x=a
}
\description{
myncurve
}
\examples{
myncurve(mu=0, sigma=1, a=2.5)
}
