% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets for Overbooking Problem}
\usage{
ntickets(N, gamma, p, n_extra = 25)
}
\arguments{
\item{N}{the number of seats on the flight}

\item{gamma}{the probability that overbooking will occur}

\item{p}{the probability that a passenger will be a "show" for the flight}

\item{n_extra}{how many extra tickets to check beyond N}
}
\value{
Named list with N, p, gamma, nd, and nc as well as two plots that show how many tickets should be sold (discrete and continuous)
}
\description{
ntickets for Overbooking Problem
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
